Draw custom focus ring using canvas
This is to let us visually overflow the canvas when dots are very near the edge. The focus ring needs to be inside a bit.

This also fixed a problem where in React dev mode with double mounting. Needed to cancel the animation request or else it could still run later with stale data.