Make auto-play more accurate
It had problems when turning Repeat on and off where it was out of sync or caused the animation to play again.

Simpler to keep the keyframe text in the DOM always up to date even if not playing, and then to use the last known keyframe text as an more accurate trigger to know when to auto play instead of after most renders.