Draw timeline from first layer
First step to adding layers of CSS properties. Timeline will now draw the active layer to interact with and some dummy background layers as a test.

The tabs will also change the current CSS prop, though they don't affect the layers yet.  Essentially, this is still the same functionality, but you can start to see where it's going.

A problem I noticed is that tabs are based on the CSS prop as their value/ID, but layers can change their CSS prop, so they are just ID'd with their index. Tabs need to be in sync with that and use the index likely.