Merge translate x/y props into one
You can only specify translate as a single entry with an x and y in the keyframe. So if one or the other is missing (and there are layers for both) we need to interpolate what the other value would be. Just leaving it "zero" or some default would move the position to the wrong place during that frame.

This needs to be done for scale x/y as well later.