Always sample at the dot positions
We know the dot positions are meaningful, so of course we should sample that point and be sure there is a keyframe for it. That's most important to be sure we hit peaks and valleys. In fact, with this we don't need nearly as many samples. For straight lines we don't need any extra samples, so we just default new layers to zero.