Draw using the current user pixel width / scale
All data now is saved in user points. Only convert to real points when actually drawing to the canvas. This means more conversions are happening, but its still fast enough.

The insets are lost with this change to simplify it, which means the borders and text at edges is clipped. Fix that next.