Vertical resize support for timeline
We need resize callbacks just like for width, and we then set the width of the user px and canvas size to match.

This also rejiggered the layout style classes for the main "rows" of the UI. These are more composable now and should work better as a separate "page" later on. The whole UI is stretchy now up to a fairly large size. Will scroll if window is too short though.