Smoother dismiss animation for notes
Was using .stack, which adds a margin-top of all non-first-child elements. That looks good, but when one is dismissed from the top, the margins then change and the notes shift a bit. The margin needs to be inside the note wrapper so that it's consistent and animates away with the wrapper. Putting on the bottom is OK since the last note doesn't have anything after it. Now margins never change during an animation.