Add working zoom buttons
Zooming is like a resize as far as rescaling the graph, so that code was pulled out from the observer to make it callable after a zoom. The resize observer is also not properly saved and discarded is redrawn.

Looking into make the actual buttons better or moving them off canvas.