Scale is not enabled properly on first load
The problem is that no activeId for the layers is set with the default layers. Getting the current layer would just default to the first, so it didn't matter then, but background layers came back wrong. That made it look the scale was in use on other tabs and would disable the group of scale props.