Move timeline buttons onto canvas on mobile
These take more space, but are simpler to understand and use. The zoom buttons are different heights, so should probably be normalized, but they don't have text and will start taking up a lot of space--not sure they fit otherwise.

Move the Copy buttons to the bottom since they are not used often, but are harder to find down there