Current tab doesn't always change with keyboard
The problem is that the saved tab ID was out of sync with the tabs, so we couldn't find the current one to find ones relative to it. Now we always save the current tab when saving the layer data, which was happening at the right time already.

Also, we only restore to the saved ID on page load, not remount, since the GlobalLayers is the source of truth once inited.